10.0 - As per my conversation with Adam on Skype, I've done 2 changes: (1) From: A common example is bus and train fares to client sites. Amounts under $40 should be reimbursed through Petty Cash. In SSW China office, amounts under ￥200 should be reimbursed through Petty Cash.
To:
A common example is bus and train fares to client sites. In Australia, amounts under $100 could be reimbursed through Petty Cash but in China, it would be around ￥300.

(2)
From:
If you’re in SSW China Office, all expense claims must:
Be work related
Only Fapiao is available
​​Be made within a month
Be authorized by the relevant Manager

To:
If you’re in China, all expense claims must:

Be work related
Only Fapiao is available
Be made within a month
Be authorized by the relevant Manager